Add service name display in open board tabs
